+++ R/dummy_imports.R
@@ -1,0 +1,3 @@
+#' @importFrom xaringan inf_mr
+#' @importFrom knitr kable
+NULL