+++ R/knitr_options.R
@@ -2,8 +2,10 @@
 #' @title Set knitr Options
 #'
 #' @description In the future, pass the options as arguments.
+#' @param page_numbers Boolean Add page numbers to the slides
+#' @param url Character (default "") optional link to package
 #' @export
-set_presentation_options = function() {
+set_presentation_options = function(page_numbers = FALSE, url = "") {
   knitr::opts_chunk$set(
     comment = "#>",
     echo = TRUE,
@@ -21,4 +23,25 @@ set_presentation_options = function() {
           dplyr.print_max = 4,
           digits = 3)
 
+  welcome_dir = system.file("example/",  package = "jrPresentation")
+  fname = file.path(welcome_dir, "style.css")
+  css = paste(readLines(fname), collapse = "\n")
+
+  if(page_numbers && nchar(url) > 0) {
+    message("Can't display both page numbers and url - pick one greedy.")
+    message("Just displaying page numbers")
+    url = ""
+  }
+
+  if(!page_numbers) {
+    css = paste(css, ".remark-slide-number {display:none;}", sep = "\n")
+  }
+  con = file("style.css", "w"); on.exit(close(con))
+  cat(css, file = con, append = FALSE)
+
+  zzz$url = url
+  invisible(NULL)
+
 }
+
+